# kvserver: emit MVCC range tombstones over rangefeeds