--- a/pkg/kv/kvserver/rangefeed/catchup_scan.go
+++ b/pkg/kv/kvserver/rangefeed/catchup_scan.go
@@ -65,6 +65,8 @@ func NewCatchUpIterator(
 	return &CatchUpIterator{
 		simpleCatchupIter: storage.NewMVCCIncrementalIterator(reader,
 			storage.MVCCIncrementalIterOptions{
+				KeyTypes:  storage.IterKeyTypePointsAndRanges,
+				StartKey:  span.Key,
 				EndKey:    span.EndKey,
 				StartTime: startTime,
 				EndTime:   hlc.MaxTimestamp,
@@ -96,6 +98,15 @@ type outputEventFn func(e *roachpb.RangeFeedEvent) error
 
 // CatchUpScan iterates over all changes in the configured key/time span, and
 // emits them as RangeFeedEvents via outputFn in chronological order.
+//
+// MVCC range tombstones are emitted at their start key, in chronological order.
+// Because the start key itself is not timestamped, these will be ordered before
+// all of the timestamped point keys that they overlap. For more details, see
+// MVCC range key info on storage.SimpleMVCCIterator.
+//
+// For example, with MVCC range tombstones [a-f)@5 and [a-f)@3 overlapping point
+// keys a@6, a@4, and b@2, the emitted order is [a-f)@3,[a-f)@5,a@4,a@6,b@2 because
+// the start key "a" is ordered before all of the timestamped point keys.
 func (i *CatchUpIterator) CatchUpScan(outputFn outputEventFn, withDiff bool) error {
 	var a bufalloc.ByteAllocator
 	// MVCCIterator will encounter historical values for each key in
@@ -103,18 +114,7 @@ func (i *CatchUpIterator) CatchUpScan(outputFn outputEventFn, withDiff bool) err
 	// events for the same key until a different key is encountered, then output
 	// the encountered values in reverse. This also allows us to buffer events
 	// as we fill in previous values.
-	var lastKey roachpb.Key
 	reorderBuf := make([]roachpb.RangeFeedEvent, 0, 5)
-	addPrevToLastEvent := func(val []byte) {
-		if l := len(reorderBuf); l > 0 {
-			if reorderBuf[l-1].Val.PrevValue.IsPresent() {
-				panic("RangeFeedValue.PrevVal unexpectedly set")
-			}
-			// TODO(sumeer): find out if it is deliberate that we are not populating
-			// PrevValue.Timestamp.
-			reorderBuf[l-1].Val.PrevValue.RawBytes = val
-		}
-	}
 
 	outputEvents := func() error {
 		for i := len(reorderBuf) - 1; i >= 0; i-- {
@@ -130,13 +130,54 @@ func (i *CatchUpIterator) CatchUpScan(outputFn outputEventFn, withDiff bool) err
 	// Iterate though all keys using Next. We want to publish all committed
 	// versions of each key that are after the registration's startTS, so we
 	// can't use NextKey.
+	var lastKey roachpb.Key
 	var meta enginepb.MVCCMetadata
+	var rangeKeysStart roachpb.Key
 	i.SeekGE(storage.MVCCKey{Key: i.span.Key})
 	for {
 		if ok, err := i.Valid(); err != nil {
 			return err
 		} else if !ok {
 			break
+		}
+
+		hasPoint, hasRange := i.HasPointAndRange()
+
+		// Emit any new MVCC range tombstones when their start key is encountered.
+		// Range keys can currently only be MVCC range tombstones.
+		//
+		// TODO(erikgrinaker): Find a faster/better way to detect range key changes
+		// that doesn't involve constant comparisons. Pebble probably already knows,
+		// we just need a way to ask it.
+		if hasRange {
+			if rangeBounds := i.RangeBounds(); !rangeBounds.Key.Equal(rangeKeysStart) {
+				rangeKeysStart = append(rangeKeysStart[:0], rangeBounds.Key...)
+
+				// Emit events for these MVCC range tombstones, in chronological order.
+				rangeKeys := i.RangeKeys()
+				for i := len(rangeKeys) - 1; i >= 0; i-- {
+					var span roachpb.Span
+					a, span.Key = a.Copy(rangeBounds.Key, 0)
+					a, span.EndKey = a.Copy(rangeBounds.EndKey, 0)
+					err := outputFn(&roachpb.RangeFeedEvent{
+						DeleteRange: &roachpb.RangeFeedDeleteRange{
+							Span:      span,
+							Timestamp: rangeKeys[i].RangeKey.Timestamp,
+						},
+					})
+					if err != nil {
+						return err
+					}
+				}
+			}
+		}
+
+		// If there's no point key here (i.e. we found a bare range key above), then
+		// step onto the next key. This may be a point key version at the same key
+		// as the range key's start bound, or a later point/range key.
+		if !hasPoint {
+			i.Next()
+			continue
 		}
 
 		unsafeKey := i.UnsafeKey()
@@ -217,9 +258,26 @@ func (i *CatchUpIterator) CatchUpScan(outputFn outputEventFn, withDiff bool) err
 			var val []byte
 			a, val = a.Copy(unsafeVal, 0)
 			if withDiff {
-				// Update the last version with its
-				// previous value (this version).
-				addPrevToLastEvent(val)
+				// Update the last version with its previous value (this version).
+				if l := len(reorderBuf) - 1; l >= 0 {
+					if reorderBuf[l].Val.PrevValue.IsPresent() {
+						return errors.AssertionFailedf("unexpected previous value %s for key %s",
+							reorderBuf[l].Val.PrevValue, key)
+					}
+					// If an MVCC range tombstone exists between this value and the next
+					// one, we don't emit the value after all -- it should be a tombstone.
+					//
+					// TODO(erikgrinaker): We can't save range keys when we detect changes
+					// to rangeKeysStart above, because NextIgnoringTime() could reveal
+					// additional MVCC range tombstones below StartTime that cover this
+					// point. We need to find a more performant way to handle this.
+					if !hasRange || !storage.HasRangeKeyBetween(
+						i.RangeKeys(), reorderBuf[l].Val.Value.Timestamp, ts) {
+						// TODO(sumeer): find out if it is deliberate that we are not populating
+						// PrevValue.Timestamp.
+						reorderBuf[l].Val.PrevValue.RawBytes = val
+					}
+				}
 			}
 
 			if !ignore {