--- a/pkg/roachpb/api.go
+++ b/pkg/roachpb/api.go
@@ -1519,6 +1519,9 @@ func (e *RangeFeedEvent) ShallowCopy() *RangeFeedEvent {
 	case *RangeFeedSSTable:
 		cpySST := *t
 		cpy.MustSetValue(&cpySST)
+	case *RangeFeedDeleteRange:
+		cpyDelRange := *t
+		cpy.MustSetValue(&cpyDelRange)
 	case *RangeFeedError:
 		cpyErr := *t
 		cpy.MustSetValue(&cpyErr)