@erikgrinaker 2022-06-10
This has a pretty hefty performance regression even in the no-range-keys case, mostly due to known performance issues with Pebble for this case. There is also some known overhead with `HasPointAndRange()`. I'm inclined to wait until the optimization work for #82045 lands, which will likely address most of this, and then address the rest of the (then hopefully minor) regression post-merge -- that PR is also required for WithDiff correctness.

```
name                                                 old time/op  new time/op  delta
CatchUpScan/mixed-case/withDiff=true/perc=0.00-24     517ms ± 1%   543ms ± 1%  +4.98%  (p=0.000 n=10+10)
CatchUpScan/mixed-case/withDiff=true/perc=50.00-24    1.03s ± 0%   1.13s ± 0%  +9.28%  (p=0.000 n=10+9)
CatchUpScan/mixed-case/withDiff=true/perc=75.00-24    901ms ± 1%   984ms ± 0%  +9.20%  (p=0.000 n=10+9)
CatchUpScan/mixed-case/withDiff=true/perc=95.00-24    186ms ± 1%   202ms ± 1%  +8.79%  (p=0.000 n=10+10)
CatchUpScan/mixed-case/withDiff=true/perc=99.00-24    157ms ± 1%   170ms ± 1%  +8.29%  (p=0.000 n=10+9)
CatchUpScan/mixed-case/withDiff=false/perc=0.00-24    771ms ± 1%   825ms ± 1%  +6.98%  (p=0.000 n=10+10)
CatchUpScan/mixed-case/withDiff=false/perc=50.00-24   659ms ± 1%   706ms ± 1%  +7.07%  (p=0.000 n=10+10)
CatchUpScan/mixed-case/withDiff=false/perc=75.00-24   590ms ± 1%   632ms ± 1%  +7.17%  (p=0.000 n=10+10)
CatchUpScan/mixed-case/withDiff=false/perc=95.00-24   170ms ± 1%   184ms ± 0%  +8.29%  (p=0.000 n=9+10)
CatchUpScan/mixed-case/withDiff=false/perc=99.00-24   153ms ± 1%   167ms ± 1%  +9.01%  (p=0.000 n=10+10)
```